requirements.txt
argparse
numpy
librosa
soundfile
praat-parselmouth
setuptools
pyworld